refactore: remove duplicate kubeconfig pkg
