+++ pkg/platform/client.go
@@ -28,8 +28,8 @@ import (
 	"github.com/loft-sh/log"
 	"github.com/loft-sh/log/survey"
 	"github.com/loft-sh/vcluster/pkg/cli/config"
-	"github.com/loft-sh/vcluster/pkg/kubeconfig"
 	"github.com/loft-sh/vcluster/pkg/platform/clihelper"
+	"github.com/loft-sh/vcluster/pkg/platform/kubeconfig"
 	"github.com/loft-sh/vcluster/pkg/projectutil"
 	perrors "github.com/pkg/errors"
 	"github.com/skratchdot/open-golang/open"
